Main Site + Cpanel
- Responsiveness of game detail pages
- lots of fixes
- similar game box in game detail
- Game submission with screenshot
- Game facts
- Scroll to top of page button

+++ Website/AtariLegend/php/admin/games/db_games_submissions.php
@@ -12,6 +12,8 @@
  *      - AL 2.0 : added messages
  *   Id: db_games_submissions.php,v 1.11 2016/08/19 STG
  *      - added change log
+ *   Id: db_games_submissions.php,v 1.12 2017/09/08 STG
+ *      - Enhanced for submissions with screenshots
  *
  ***************************************************************************/
 
@@ -50,6 +52,33 @@ if ($action == "delete_submission") {
     //****************************************************************************************
 
     if (isset($submit_id)) {
+        
+        //Let's first check if this submission has screenshots.
+        $query_submit_screenshot = $mysqli->query("SELECT * FROM screenshot_game_submitinfo WHERE game_submitinfo_id = " . $submit_id . "") or die("something is wrong with mysqli of submissions screenshots");
+        
+        while ($sql_submit_screenshot = $query_submit_screenshot->fetch_array(MYSQLI_BOTH))
+        {
+            $screenshot_id = $sql_submit_screenshot['screenshot_id'];
+
+            //get the extension
+            $SCREENSHOT = $mysqli->query("SELECT * FROM screenshot_main
+                                      WHERE screenshot_id = '$screenshot_id'") or die("Database error - selecting screenshots");
+
+            $screenshotrow  = $SCREENSHOT->fetch_array(MYSQLI_BOTH);
+            $screenshot_ext = $screenshotrow['imgext'];
+
+            $sql = $mysqli->query("DELETE FROM screenshot_main WHERE screenshot_id = '$screenshot_id' ");
+            $sql = $mysqli->query("DELETE FROM screenshot_game_submitinfo WHERE screenshot_id = '$screenshot_id' ");
+
+            $new_path = $game_submit_screenshot_save_path
+            ;
+            $new_path .= $screenshot_id;
+            $new_path .= ".";
+            $new_path .= $screenshot_ext;
+
+            unlink("$new_path");  
+        }
+                 
         create_log_entry('Games', $submit_id, 'Submission', $submit_id, 'Delete', $_SESSION['user_id']);
 
         $sql = $mysqli->query("DELETE FROM game_submitinfo WHERE game_submitinfo_id = '$submit_id'") or die("couldn't delete game_submissions quote");
@@ -69,29 +98,40 @@ if ($action == "move_submission_tocomment") {
     //****************************************************************************************
     // This is the move to comments place
     //****************************************************************************************
-
+    
     if (isset($submit_id)) {
-        $query_submit = $mysqli->query("SELECT * FROM game_submitinfo WHERE game_submitinfo_id = " . $submit_id . "") or die("something is wrong with mysqli");
-        $sql_submit = $query_submit->fetch_array(MYSQLI_BOTH) or die("something is wrong with mysqli2");
+        //Let's first check if this submission has screenshots. If so, it is not suited for movement to comment section!
+        $query_submit_screenshot = $mysqli->query("SELECT * FROM screenshot_game_submitinfo WHERE game_submitinfo_id = " . $submit_id . "") or die("something is wrong with mysqli of submissions screenshots");
+        $v_rows = $query_submit_screenshot->num_rows;
+        
+        if ($v_rows > 0)
+        {
+            $_SESSION['edit_message'] = "Submission has screenshots and is not suited for the comment section";
+        }
+        else
+        {
+            $query_submit = $mysqli->query("SELECT * FROM game_submitinfo WHERE game_submitinfo_id = " . $submit_id . "") or die("something is wrong with mysqli");
+            $sql_submit = $query_submit->fetch_array(MYSQLI_BOTH) or die("something is wrong with mysqli2");
 
-        $submit_text   = $sql_submit['submit_text'];
-        $submit_text   = $mysqli->real_escape_string($submit_text);
-        $sub_timestamp = $sql_submit['timestamp'];
-        $sub_user_id   = $sql_submit['user_id'];
-        $sub_game_id   = $sql_submit['game_id'];
+            $submit_text   = $sql_submit['submit_text'];
+            $submit_text   = $mysqli->real_escape_string($submit_text);
+            $sub_timestamp = $sql_submit['timestamp'];
+            $sub_user_id   = $sql_submit['user_id'];
+            $sub_game_id   = $sql_submit['game_id'];
 
-        $sql = $mysqli->query("INSERT INTO comments (comment,timestamp,user_id) VALUES ('$submit_text','$sub_timestamp','$sub_user_id')") or die("something is wrong with INSERT mysql3");
+            $sql = $mysqli->query("INSERT INTO comments (comment,timestamp,user_id) VALUES ('$submit_text','$sub_timestamp','$sub_user_id')") or die("something is wrong with INSERT mysql3");
 
-        $new_comment_id = $mysqli->insert_id;
+            $new_comment_id = $mysqli->insert_id;
 
-        $sql = $mysqli->query("INSERT INTO game_user_comments (game_id,comment_id) VALUES ('$sub_game_id',LAST_INSERT_ID())") or die("something is wrong with INSERT mysql4");
+            $sql = $mysqli->query("INSERT INTO game_user_comments (game_id,comment_id) VALUES ('$sub_game_id',LAST_INSERT_ID())") or die("something is wrong with INSERT mysql4");
 
-        $sql = $mysqli->query("DELETE FROM game_submitinfo WHERE game_submitinfo_id = " . $submit_id . "") or die("couldn't delete game_submissions quote");
+            $sql = $mysqli->query("DELETE FROM game_submitinfo WHERE game_submitinfo_id = " . $submit_id . "") or die("couldn't delete game_submissions quote");
+            
+            $_SESSION['edit_message'] = "Submission converted to game comment";
+
+            create_log_entry('Games', $new_comment_id, 'Comment', $new_comment_id, 'Insert', $_SESSION['user_id']);
+        }
     }
-
-    $_SESSION['edit_message'] = "Submission converted to game comment";
-
-    create_log_entry('Games', $new_comment_id, 'Comment', $new_comment_id, 'Insert', $_SESSION['user_id']);
 
     if ($list == "done") {
         header("Location: ../games/submission_games.php?v_counter=$v_counter&list=$list");
